파일 이동 및 삭제
