--- Assets/LTH/LTH_Prefabs/Player.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 28f113d40e8a8c841a3fc4234a67491e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
