--- Assets/LTH/Scripts/RoundUIController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e137b04ca12fe75499887d4d332de460
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
