--- Assets/LTH/Scripts/TurnUIController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4d36693951eef8d47a306723ead8e6c7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
